has all possible moves, each ranked, and outputs a moves file. Not yet on multiple ('?',3,'Y',red)'s
The moves file has the format of "(3,4,1),(2,5,1) etc (there is only a comma between entries no comma on the end). If more is needed please let me know. Put into the same directory as the program and is named "Paths.txt" - there's nothing fancy for protecting already written files - --- yet. 
Also the findNeighbors function was previously limited in the number of possible moves. It now creates an optimal move, then considers all other moves - which probably makes the optimal move redundant, but I'll play with it. It's still pretty fast.