with a simple getNeighbors function
This version is the truer version of what was originally done. It calculates the optimal move and then ranks and puts the 10 other very simple moves into the moves queue. The moves are all just in one dimension - like just x, just y, just z. So it takes longer to get to the target, but it actually works without the optimal first move unlike the first one I did.