Replace the method processList() with this
It bugged me that the sim didn't process a list that might be out of order  - in terms of the Y values of the cubes being non-sequential and forming a tower.
For example if any list were: 0 6 0 purple, 0 3 0 yellow, 0 4 0 red, 0 5 0 green, 0 1 0 white,  0  2  0   blue, 0 0 0 blue, 10 10 10 drone
It would be flagged as having floaters and fail. This new processList() takes care of that. 
It should start at line 227 in the sim code - so replace the entire method with this new code.